Move generate code to a service file
